+++ Managers/UserManager.class.php
@@ -13,6 +13,11 @@ class UserManager extends AbstractConnectionManager{
 	
 	function signIn($name, $password)
 	{
+		if($name == "" || $password == "")
+		{
+	        return false;
+		}
+
 		$password = hash('sha512', $password, false);
 
 		$req = $this->db->prepare('SELECT idUser FROM User WHERE name = :name AND password = :password');
@@ -33,8 +38,10 @@ class UserManager extends AbstractConnectionManager{
 
 	function signUp($name, $mail, $password, $adminLevel)
 	{
-	    if($name == "" || $email == "" || $password == "")
+		if($name == "" || $email == "" || $password == "")
+		{
 	        return false;
+		}
 	        
 		$password = hash('sha512', $password, false);
         
@@ -58,6 +65,11 @@ class UserManager extends AbstractConnectionManager{
 
 	function deleteUser($name)
     {
+		if($name == "")
+		{
+	        return false;
+		}
+
         $req = $this->db->prepare('DELETE FROM User WHERE name = :name');
         $req->bindValue(':name', $name, PDO::PARAM_STR);
         $req->execute();
@@ -75,6 +87,11 @@ class UserManager extends AbstractConnectionManager{
 
 	function createOrganisation($nameUser, $nameEvent, $isCreator)
 	{
+		if($nameUser == "" || $nameEvent == "" || $isCreator == "")
+		{
+	        return false;
+		}
+
 		$req = $this->db->prepare('SELECT idUser FROM User WHERE name = :name');
         $req->bindValue(':name', $nameUser, PDO::PARAM_STR);
         $req->execute();
